Add working version of source command.
This version is working, but not well tested. I need to do some more tests and go through the logic again to make sure I do not have anything redundant leftover. I had to do some path hackery due to how the bot is installed as package on the docker image, this may or may not be compatible with the deploy environment, so we may need to change some path logic or change the base path to be an environment variable.